deploy-strapi - fix waiting for production
Wait for production backend to be up also when using tags.
Turns out github.ref_name will be v1.2.3 on a tag, so we cannot easily
check if this tag is on master branch or not. But lets assume all v1.2.3
tags are on production, as it would be a mistake to do otherwise. I'm
admitting this is brittle. Why GH actions does not offer a branch name
variable? why?